Move dss.storage.validator to dss.index.es.validator
